+++ src/documents.ts
@@ -69,6 +69,15 @@ export type Document<T extends object = any> = T &
 export type Edge<T extends object = any> = T & DocumentMetadata & EdgeMetadata;
 
 /**
+ * Type representing patch data for a given object type to represent a payload
+ * ArangoDB can apply in a document PATCH request (i.e. a partial update).
+ *
+ * This differs from `Partial` in that it also applies itself to any nested
+ * objects recursively.
+ */
+export type Patch<T = object> = { [K in keyof T]?: T[K] | Patch<T[K]> };
+
+/**
  * An object with an ArangoDB document `_id` property.
  *
  * See {@link DocumentMetadata}.